+++ src/components/micro/input-short-text/styled-input-group.js
