+++ src/components/micro/input-short-text/styled-label.js
